add initial implementation and API
This brings in the initial implementation that is being extracted
from other projects where I have repeated this pattern/API
multiple times.

This version of the pattern has been tweaked to not only work
nicely with MuchRails, but also to be used in the context of an
external gem (with regards to configuration and setup).